# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GAC授权中心",
        "title": "GAC授权中心",
        "termsOfService": "https://paymedia.yuque.com/os/hb9uhc",
        "contact": {
            "name": "org.pm.gac.service",
            "url": "https://www.apifox.cn/web/main"
        },
        "license": {
            "name": "GIN.swagger",
            "url": "https://github.com/swaggo/swag/blob/master/README_zh-CN.md#%E5%A6%82%E4%BD%95%E4%B8%8Egin%E9%9B%86%E6%88%90"
        },
        "version": "2.0"
    },
    "host": "api.t.dev.pay.fun",
    "basePath": "/gac/v1",
    "paths": {
...
        "/admin/log/action_option": {
            "get": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GAC/日志管理"
                ],
                "summary": "行为下拉",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/adminApi.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/assembler.Action"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    }
                }
            }
        },
        "/admin/log/module_option": {
            "get": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GAC/日志管理"
                ],
                "summary": "模块下拉",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "description": "The object content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.OrgSearchReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/adminApi.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/assembler.Action"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    }
                }
            }
        },
        "/admin/log/page": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GAC/日志管理"
                ],
                "summary": "日志列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "description": "The object content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.LogSearchReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/adminApi.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/assembler.AdminLogAssembler"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "adminApi.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "assembler.Action": {
            "type": "object"
        },
...
        "assembler.AdminLogAssembler": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "行为",
                    "type": "string"
                },
                "action_user": {
                    "description": "操作人",
                    "type": "string"
                },
                "app_pk": {
                    "description": "应用",
                    "type": "integer"
                },
                "created_ip": {
                    "description": "ip",
                    "type": "string"
                },
                "created_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "module": {
                    "description": "模块",
                    "type": "string"
                },
                "org_pk": {
                    "description": "组织",
                    "type": "integer"
                },
                "pk": {
                    "type": "integer"
                },
                "record": {
                    "description": "记录",
                    "type": "string"
                }
            }
        },
...
        "req.LogSearchReq": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "action_user": {
                    "type": "string"
                },
                "app_pk": {
                    "type": "integer"
                },
                "current": {
                    "description": "当前页",
                    "type": "integer"
                },
                "module": {
                    "type": "string"
                },
                "org_pk": {
                    "type": "integer"
                },
                "page_size": {
                    "description": "页大小",
                    "type": "integer"
                },
                "pagination": {
                    "description": "是否使用分页查询",
                    "type": "boolean"
                }
            }
        },
...
        "req.OrgSearchReq": {
            "type": "object",
            "properties": {
                "current": {
                    "description": "当前页",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "page_size": {
                    "description": "页大小",
                    "type": "integer"
                },
                "pagination": {
                    "description": "是否使用分页查询",
                    "type": "boolean"
                },
                "pk": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    }
}